+++ CHANGES.md
@@ -11,8 +11,9 @@ Fast and flexible PHP library rendering plain ASCII tables.
   * Added option to configure column default align.
   * Added option to separately control table column title align.
   * Custom column max width is now correctly handled.
-  * Too long column labels values are now truncated with ellipsis.
-  * more tests.
+  * Fixed width columns are now truncated automatically with ellipsis.
+  * Improved code creating row cells from plain string arrays. 
+  * Added more tests.
 
 * v1.1.0 (2022-10-17)
   * Corrected `\ArrayAccess` implementation.